Extract validation indicator to separate file
