+++ packages/components/src/validated-form-controls/control-with-error.tsx
@@ -1,13 +1,7 @@
 /**
- * External dependencies
- */
-import clsx from 'clsx';
-
-/**
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { error, published } from '@wordpress/icons';
 
 /**
  * External dependencies
@@ -24,9 +18,7 @@ import {
  */
 import { withIgnoreIMEEvents } from '../utils/with-ignore-ime-events';
 import type { ValidatedControlProps } from './components/types';
-
-import Icon from '../icon';
-import Spinner from '../spinner';
+import { ValidationIndicator } from './validation-indicator';
 
 function appendRequiredIndicator(
 	label: React.ReactNode,
@@ -48,39 +40,6 @@ function appendRequiredIndicator(
 		);
 	}
 	return label;
-}
-
-function ValidationIndicator( {
-	type,
-	message,
-}: {
-	type: 'validating' | 'valid' | 'invalid';
-	message?: string;
-} ) {
-	const ICON = {
-		valid: published,
-		invalid: error,
-	};
-	return (
-		<p
-			className={ clsx(
-				'components-validated-control__status',
-				`is-${ type }`
-			) }
-		>
-			{ type === 'validating' ? (
-				<Spinner className="components-validated-control__status-spinner" />
-			) : (
-				<Icon
-					className="components-validated-control__status-icon"
-					icon={ ICON[ type ] }
-					size={ 16 }
-					fill="currentColor"
-				/>
-			) }
-			{ message }
-		</p>
-	);
 }
 
 /**
